COMMIT V1.O

+++ app/src/main/java/com/example/entertainer/data/Movie.kt
@@ -12,7 +12,7 @@ data class Movie(
     val title: String,
     val genre: MovieCategories,
     val director: String,
-    val duration: Int,
+    val duration: Double,
     val actors: String,
     val rating: String,
     val image: ByteArray?
